// src/types/carga.ts
export interface IObjeto {
  etiqueta: string;
  pacotePostagem: string;
  clienteId: number;
  agfId: string;
  departamento: string;
  os: string;
  nomeRemetente: string;
  enderecoRemetente: string;
  numeroRemetente: string;
  complementoRemetente: string;
  bairroRemetente: string;
  cidadeRemetente: string;
  ufRemetente: string;
  cepRemetente: string;
  nomeServico: string;
  produto: string;
}